moved preprocessing for dataset
